+++ src/rinq/amqp/internal/commandamqp/factory.go
@@ -37,6 +37,7 @@ func New(
 		queues,
 		channels,
 		cfg.Logger,
+		cfg.Tracer,
 	)
 	if err != nil {
 		return nil, nil, err
@@ -49,6 +50,7 @@ func New(
 		queues,
 		channels,
 		cfg.Logger,
+		cfg.Tracer,
 	)
 	if err != nil {
 		invoker.Stop()