+++ src/rinq/amqp/internal/notifyamqp/message.go
@@ -4,8 +4,12 @@ import (
 	"errors"
 	"fmt"
 
+	opentracing "github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
 	"github.com/rinq/rinq-go/src/rinq"
+	"github.com/rinq/rinq-go/src/rinq/amqp/internal/amqputil"
 	"github.com/rinq/rinq-go/src/rinq/ident"
+	"github.com/rinq/rinq-go/src/rinq/internal/traceutil"
 	"github.com/streadway/amqp"
 )
 
@@ -99,4 +103,19 @@ func unpackConstraint(msg *amqp.Delivery) (rinq.Constraint, error) {
 	}
 
 	return nil, errors.New("constraint header is not a table")
+}
+
+func unpackSpanOptions(msg *amqp.Delivery, t opentracing.Tracer) (opts []opentracing.StartSpanOption, err error) {
+	sc, err := amqputil.UnpackSpanContext(msg, t)
+
+	if err == nil {
+		opts = append(opts, traceutil.CommonSpanOptions...)
+		opts = append(opts, ext.SpanKindConsumer)
+
+		if sc != nil {
+			opts = append(opts, opentracing.FollowsFrom(sc))
+		}
+	}
+
+	return
 }