+++ src/rinq/internal/localsession/session_logging.go
@@ -11,6 +11,16 @@ import (
 	"github.com/rinq/rinq-go/src/rinq/trace"
 )
 
+func logCreated(
+	logger rinq.Logger,
+	ref ident.Ref,
+) {
+	logger.Log(
+		"%s session created",
+		ref.ShortString(),
+	)
+}
+
 func logCall(
 	logger rinq.Logger,
 	msgID ident.MessageID,
@@ -79,8 +89,13 @@ func logAsyncRequest(
 	ns string,
 	cmd string,
 	out *rinq.Payload,
+	err error,
 	traceID string,
 ) {
+	if err != nil {
+		return // request never sent
+	}
+
 	logger.Log(
 		"%s called '%s::%s' command asynchronously (%d/o) [%s]",
 		msgID.ShortString(),